requirements.txt
Django==4.0.1
djangorestframework==3.13.1
django-filter==21.1
factory-boy==3.2.1
Faker==11.3.0
django-debug-toolbar==3.2.4